From  Edward Sheldrake: Wireshark crashes if you open the file import dialog and choose something from the "Encapsulation type" list.
https://bugs.wireshark.org/bugzilla/show_bug.cgi?id=6729

svn path=/trunk/; revision=40448

+++ gtk/file_import_dlg.c
@@ -203,7 +203,7 @@ encap_co_changed(GtkComboBox *widget, gpointer data)
 
     if (result)
     {
-        guint8 encap;
+        guint encap;
         GtkTreeModel *model = gtk_combo_box_get_model(widget);
         gtk_tree_model_get(model, &iter, 1, &encap, -1);
